# services/booking-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para reservas, asientos, checkout y tickets.",
        "title": "SeatGuard Booking Service API",
        "contact": {
            "name": "Lucas Cabral",
            "email": "lucassimple@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/tickets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene todos los tickets (solo admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Obtener todos los tickets",
                "responses": {
                    "200": {
                        "description": "Lista de tickets",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un ticket a partir de un order completado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Crear ticket desde endpoint",
                "parameters": [
                    {
                        "description": "Datos del ticket",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket creado",
                        "schema": {
                            "$ref": "#/definitions/models.TicketPDF"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Order no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Ticket ya existe",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/by-id/{ticketID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un ticket específico por su ID (solo owner o admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Obtener ticket por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del ticket",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.TicketPDF"
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/{orderID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene los datos del ticket sin el PDF binario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Obtener metadata del ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del order",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadata del ticket",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un ticket (soft delete) - solo owner o admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Eliminar ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del order",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket eliminado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/{orderID}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Genera y descarga el PDF del ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Descargar PDF del ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del order",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF del ticket",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tickets/{orderID}/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Regenera el PDF del ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Regenerar PDF del ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del order",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF del ticket",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Seat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventId": {
                    "description": "Relaciones",
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lockedAt": {
                    "description": "Cuando se desbloquea",
                    "type": "string"
                },
                "lockedBy": {
                    "description": "Control de Bloqueo Temporal",
                    "type": "string"
                },
                "number": {
                    "description": "\"10\", \"A1\", etc.",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "section": {
                    "description": "Identificación del asiento (ej: A-10, B-5)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado actual del asiento\nenums: AVAILABLE, RESERVED, SOLD, BLOCKED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SeatStatus"
                        }
                    ]
                },
                "ticketId": {
                    "description": "ID del ticket final si se vende",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SeatStatus": {
            "description": "Estado del asiento",
            "type": "string",
            "enum": [
                "AVAILABLE",
                "LOCKED",
                "SOLD"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusLocked",
                "StatusSold"
            ]
        },
        "models.TicketPDF": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                },
                "name": {
                    "type": "string"
                },
                "orderId": {
                    "type": "string"
                },
                "paymentIntentId": {
                    "type": "string"
                },
                "paymentProvider": {
                    "type": "string"
                },
                "pdfGeneratedAt": {
                    "type": "string"
                },
                "pdfVersion": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Formato: Bearer \u003ctoken\u003e\nOperaciones para gestionar eventos\nOperaciones para gestionar asientos\nOperaciones para gestionar órdenes de reserva\nOperaciones para gestionar el proceso de checkout",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}